Change form reading one system to reading multiple

+++ src/planet_system/center_star.rs
@@ -3,7 +3,7 @@ pub struct CenterStar {
     pub name: String,
     pub mass: f32,
     pub radius: f32,
-    pub effective_temperature: u16,
+    pub effective_temperature: f32,
 }
 
 impl CenterStar {
@@ -12,7 +12,7 @@ impl CenterStar {
             name: "".to_string(),
             mass: 0.0,
             radius: 0.0,
-            effective_temperature: 0,
+            effective_temperature: 0.0,
         }
     }
 }
@@ -23,7 +23,7 @@ impl From<Vec<String>> for CenterStar {
             name: value[0].to_string(),
             mass: value[1].parse::<f32>().unwrap_or_default(),
             radius: value[2].parse::<f32>().unwrap_or_default(),
-            effective_temperature: value[8].parse::<u16>().unwrap_or_default(),
+            effective_temperature: value[8].parse::<f32>().unwrap_or_default(),
         }
     }
 }
