+++ src/planet_system/planet.rs
@@ -7,7 +7,7 @@ pub struct Planet {
     pub radius: f32,
     pub semi_major_axis: f32,
     pub eccentricity: f32,
-    pub orbital_period: u32,
+    pub orbital_period: f32,
     pub moons: Vec<Moon>
 }
 
@@ -19,7 +19,7 @@ impl From<Vec<String>> for Planet {
             radius: value[2].parse::<f32>().unwrap_or_default(),
             semi_major_axis: value[3].parse::<f32>().unwrap_or_default(),
             eccentricity: value[4].parse::<f32>().unwrap_or_default(),
-            orbital_period: value[5].parse::<u32>().unwrap_or_default(),
+            orbital_period: value[5].parse::<f32>().unwrap_or_default(),
             moons: vec![],
         }
     }
